Supporting the 19.1.10 version, demo refresh release.
配套19.1.10版本，Demo刷新发布，主要更新包括：
1. 解决部分问题；
2. 新增修改密码功能；
3. 新增支持新会议AS环境匿名会议功能。

Supporting the 19.1.10 version, demo refresh release. The main updates are as follows:
1. Resolve some problems.
2. SDK new support app modifying password function.
3. SDK new support app joining anonymous conference function on the new conference AS environment.

+++ sample/EC_SDK_Demo/Demo/app/src/main/java/com/huawei/opensdk/ec_sdk_demo/logic/call/VideoCallPresenter.java
@@ -27,7 +27,7 @@ public class VideoCallPresenter extends MVPBasePresenter<IVideoCallContract.Vide
     private final Context mContext;
     private final CallFunc mCallFunc;
 
-    private int mCallID;
+    private long mCallID;
 
     private int mAudioRoute;
     private int mCallType;
